[pinpoint] Separate Execution exceptions from result_values.
It's not good practice to mix types.

Having separate exception and result_values will make the UI easier, because it
don't have to guess whether result_values contains numeric data, or exception
strings, or a mixture of all 3. And it'll be easier to associate which results
go with which swarming task.

Execution.exception will always be a basestring, and Execution.result_values
will always be a tuple of numbers.

Review-Url: https://chromiumcodereview.appspot.com/3008183002