[Tracing] Add add_shared_diagnostics.AddValueDiagnostics
In preparation for adding tracing/bin/add_shared_diagnostics, which we intend to
support adding simple diagnostics directly, this CL adds a function
add_shared_diagnostics.AddValueDiagnostics.

In so doing, it pluralizes the AddSharedDiagnostics method to avoid
deserialization overhead, and moves away from using a JavaScript implementation
for simplicity's sake.

BUG=catapult:#3763

Review-Url: https://chromiumcodereview.appspot.com/3000643002