[Telemetry]  Let --also-run-disabled-tests override StoryExpectations.DisableBenchmark
Enable benchmarks disabled with StoryExpectations.DisableBenchmark to run when --also-run-disabled-tests flag is set.

As a holdover from when it was PermanentlyDisableBenchmark, we currently have it so that the disabling cannot be overridden with --also-run-disabled-tests. Since we now use SUPPORTED_PLATFORMS to indicate a benchmark cannot run on a given platform, and DisableBenchmark is for temporary disablings, it makes sense to make this overridable now.

BUG=catapult:#3896

Review-Url: https://chromiumcodereview.appspot.com/3020443002