Add tab.StopAllServiceWorkers()
This CL adds a method StopAllServiceWorkers() to tab.py .
In order to implement this method, I also made InspectorServiceWorker class and use websocket connections logic with devtool ServiceWorker domain in it.

ServiceWorker backend is enabled for all InspectorBackends on initializing InspectorServiceWorker (with devtools method ServiceWorker.enable).

Design doc of perf benchmark for PWA:
https://docs.google.com/document/d/1Nf97CVp1X7aSqvAspyJ7yOCDyr1osUNrnfrGwZ_Yuuo/edit?usp=sharing

BUG=chromium:736697

Review-Url: https://chromiumcodereview.appspot.com/3018603002