Fix telemetry's IsBrowserRunning check
This method currently checks to see if there's a process that matches
the _prefix_ 'org.chromium.chrome' and returns true if there is.
However, even if the browser itself is shutdown, there may be other
processes running that match that prefix which can confuse telemetry.

This patch fixes the issue by using GetApplicationPids which uses an
exact match rather than a prefix.

BUG=catapult:#3912

Review-Url: https://chromiumcodereview.appspot.com/3018503002