Fix uses of /deep/ in trace viewer.
There are several old modules that use tr.ui.b.define() and /deep/ CSS rules.
Support for /deep/ is being removed from Chrome in M63, which will branch in a few
weeks.
This CL moves all of the /deep/ CSS rules into javascript.
This is not the best way to solve this problem.
The best way to solve this problem would be to turn all of these old modules into
proper Polymer dom-modules: https://github.com/catapult-project/catapult/issues/3058
However, that will be tricky and take time that would be better spent on other
things. This CL is a stop-gap.

ListAndAssociatedView appears entirely unused, so it is removed.

http://www/~benjhayden/3017523002.html

https://github.com/catapult-project/catapult/issues/1691
Bug: chromium:765447

Change-Id: I505ccaf708dd7f37af129a8e4657f9a5d812b3f2
Reviewed-on: https://chromium-review.googlesource.com/689742
Commit-Queue: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Fadi Meawad <fmeawad@chromium.org>