Fix telemetry gesture speeds
In b0d47f5f4d4b9717bfdd43ae6d40257235d0848f I made it so telemetry JS
scripts pass values to gpuBenchmarking in the coordinate space of the
(visual) viewport. This missed the speed value which is assumed to be
in CSS pixels/sec. Not performing this adjustment meant that the max
scrollable distance calculation based on speed and maximum time was
wrong.

Bug: chromium:772447
Change-Id: Ifdf56781f90b8efabfec5f9e72dea8b125db7bfd
Reviewed-on: https://chromium-review.googlesource.com/733980
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>