Replace RelatedHistogramBreakdown in longTasksMetric with RelatedNameMap.
longTasksMetric is not used by any benchmarks, so it's ok to change its
Histogram names. longTasksMetric is still useful in the metrics side panel as a
trace viewer feature, and as an example metric.

Demo: http://www/~benjhayden/764877.html

This is part of simplifying TBMv2 relationship diagnostics:
https://docs.google.com/document/d/1Lv-KoSFEq4EvWoXCTk_TO-YVqPWg53RwE4x1xBR0QoI/edit

Bug: catapult:#3761
Change-Id: I0492ad10639933b49d0f9303cc611af3622a9f4f
Reviewed-on: https://chromium-review.googlesource.com/764877
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>