Add landing page to md-pages branch
The md-pages branch is served on catapult.gsrc.io, and we don't have a
good way to automatically keep this branch up-to-date right now. Until we
do, this CL leaves a landing pages that points to googlesource.

Please note that the bug line below is only to satisfy the out-of-date
PRESUBMIT.py on the branch.

BUG=catapult:#1234
Change-Id: I7a09922cde0b081cfb4f8a73e867bddfbb8ff3c0
Reviewed-on: https://chromium-review.googlesource.com/772846
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>