Fix "View Options" dropdown in trace viewer.
The tr-ui-b-dropdown dom-module encapsulates a modal dialog that drops down from
a button. When the user clicks outside of the modal dialog, it closes. When the
user clicks inside of the modal dialog, it should not close. Due to the way that
modal dialogs are implemented, clicking inside of the dialog produces the same
type of event as clicking outside of it, so it is common for modules such as
tr-ui-b-dropdown to compare the coordinates of the click event to the bounding
rect of the dialog in order to determine whether the click was inside or outside
of the dialog.

Polymer defines 'tap' events as a generalization of click events that support
touch screens as well as mouse pointers.  However, there is a crucial difference
between click events and polymer tap events: click event coordinates are given
by event.clientX and event.clientY, whereas tap event coordinates are given by
event.detail.x and event.detail.y.

When a recent change upgraded the click event handler to a tap event handler, it
failed to update the clientX/clientY references to point to the Polymer tap
event's detail.x/y.

This CL fixes the click-event-style clientX/clientY references to correctly
reference the tap-event-style detail.x/y. The new test fails without the fix.

Bug: chromium:785138
Change-Id: I006652066a2d56a6219757f1815bfe9bad7bb54e
Reviewed-on: https://chromium-review.googlesource.com/795110
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>