[TBMv2] Save exception thrown by metric
Currently, if a metric throws an exception, we re-raise it. However, we want
to allow metrics that are used for correctness testing (e.g.
leakDetectionMetric) to be able to throw to indicate correctness failure,
while still outputting histograms.

This CL adds a try/catch block to allow an exception raised while running
metrics to be translated into an MRE Failure, which will ensure that we don't
bail out before we're ready to emit histograms.

Change-Id: I7b09c4fb208431322c844b6c02458ed8d4917e4b
Reviewed-on: https://chromium-review.googlesource.com/768530
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Ethan Kuefner <eakuefner@chromium.org>