[Telemetry] Add test Expectation file parser for TA/DA telemetry integration.
This uses the format for tagged tests described in go/chromium-test-list-format.
It parses the file and returns a list of dictionaries that contain:
{
  'reason': crbug.com123/Bug(),
  'test': test_name,
  'conditions': [list, of, conditions]
  'results': [list, of, results]
}

The only enteries that are guarenteed to not be empty by the defined grammar are
test and results. Reason and conditions will be None and [] if they are not
included in the expectation entry.

The next CL will use this list to create StoryExpectations that are already used
by telemetry for disabling stories.

Bug: chromium:781409
Change-Id: I1979fa19c56afd86924beb24daf9fa6eaca673f4
Reviewed-on: https://chromium-review.googlesource.com/772805
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Commit-Queue: rnephew <rnephew@chromium.org>