[Telemetry] Fix browser_backend.pid for Android
The old code used to expect finding a single process matching the name
of the browser app.

This, however, is not always reliable in cases where the browser
forks its main process and, for a short period of time, two processes
may be found with the same name.

Instead we now look for the single process with the matching name
which is also a direct descendant of the Android zygote.

Bug: chromium:794178
Change-Id: Ie577e19e6b63bcb5192fbe16ef79b7cbbe09387e
Reviewed-on: https://chromium-review.googlesource.com/822261
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>