[Telemetry] Move computation of chrome_startup_flags to own module
For now the new module computes the list of startup args that depend
solely on the browser_options object, and which can be computed
without reference to anything else.

The computation of startup args is now started by possible-browsers,
*before* the brower backend actually exits. This is needed on Android
to decouple setting the chrome command line from starting the browser.

Follow up CLs will migrate the rest of args, still computed on
browser-backends, to be computed on the possible-browsers instead.

Bug: chromium:787834
Change-Id: I856c9bc56c7cc3447eb9f04e6582278c9dfc06dc
Reviewed-on: https://chromium-review.googlesource.com/854054
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>