Log wmic exit code when it fails
Previously, when wmic failed, I tried to log the output that it gave,
hoping that it would give some clue as to why it failed. Unfortunately,
it looks like when it fails, _no_ output is given.

wmic does give very specific exit codes when it fails, though, so I'm
hoping that this latest logging will help us diagnose the failure.

TBR=nednguyen@google.com
Bug: chromium:803210
Change-Id: I9f31a08712545940c4c9f2a82aac77fb99fb013e
Reviewed-on: https://chromium-review.googlesource.com/874599
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>