[pinpoint] Fix retrieval of isolated JSONs.
Isolate server can return file contents in two formats. If the file is short, it
returns it directly in the 'content' field. If it's long, it gives a url to the
file in Cloud Storage. The isolate_service module incorrectly treated the first
as JSON and the second as plaintext. They should both be treated as text of an
unknown format.

Bug: catapult:#4190
Change-Id: I44fdfc759a390f81198fb1e22d97fc8b2a18de64
Reviewed-on: https://chromium-review.googlesource.com/882310
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: David Tu <dtu@chromium.org>