[Dashboard] Don't create extra statistics rows for legacy tests/memory metrics
This CL allows us to achieve test hierarchy parity with the /add_point
pipeline. In particular, for most types of tests, we want to create the same
suffixed statistics rows that /add_point does from chart JSON. There are two
exceptions to this: memory metrics, for which we want to only produce _avg,
and legacy tests, for which we don't want to produce these suffixed rows at
all.

Fixes #4201. Also addresses #4203 but can't mention only one bug in the Bug:
line, which appears to be a depot_tools bug. I'll file a bug about that later.

Bug: catapult:#4203
Change-Id: Ic258a43bc1482a205b4cac4fa3f306f55488b4c4
Reviewed-on: https://chromium-review.googlesource.com/892085
Commit-Queue: Ethan Kuefner <eakuefner@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>