[results.html] Show All by default.
Currently, Show All is unchecked by default in order to improve loading
times and not overwhelm the user with too much data.
However, it is easy for users to not see the checkbox and incorrectly
conclude that the measurement they're seeking is missing.
This happened recently in http://crbug.com/814678

This CL checks Show All by default in order to help users find
measurements. It can cause a moment of the sense of a firehose, but the
search box is discoverable enough.

Change-Id: Ida2d04b87d51dbf1cd88ef462bcc0bd14a429094
Reviewed-on: https://chromium-review.googlesource.com/949564
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>