More precise self time calculation
We have had multiple instances of negative values in breakdowns, and the
root cause was imprecise self time calculations. This CL makes self time
calculation more precise when a slice is partially contained in
rangeOfInterest.

This CL also switches to using wall time range of interest for CPU time
ranges, because CPU time range of interest is not always available (for
example, when the user provides it using the range tool), and it was
easy to incorrectly use wall time range of interest and get very wrong
data.

Bug: catapult:#4214

Change-Id: I9e30dedd235dd95c0a8aabee40a68ef320247039
Reviewed-on: https://chromium-review.googlesource.com/939481
Commit-Queue: Deepanjan Roy <dproy@chromium.org>
Reviewed-by: Liquan Gu <maxlg@chromium.org>
Reviewed-by: Fadi Meawad <fmeawad@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>