Reland "[Telemetry] Move apk finding code to Android backend settings"
A first step towards encapsulating the logic to find and install apks
for individual Android browser variants, this CL introduces a
backend_settings.FindLocalApk() method.

As a side effect of this change the logic of the corresponding browser
finder becomes cleaner and now, even if --chrome-root is not given,
allows to select any known browser which is already installed on the
device.

We also take the chance to make _HaveLocalAPK and
_HaveWebViewEmbedderAPK private, as the logic to handle those will
be eventually moved to the banckend_settings as well.

Fixed bug from previous attempt to land where the platform_backend
mock had a spec of the wrong object.

Bug: chromium:815133
Change-Id: Ia4c5a5ee65a7ac84fd7b94d440a05001ef396ef5
Reviewed-on: https://chromium-review.googlesource.com/952907
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>