Fix index of timeseries data frames
A previous CL introduced a new column in the timeseries data frame,
however this inadvertently broke the set of columns that are regarded
as "index" for the database table (causing point_id to be excluded).

Fix bug and tweak one of the tests so this would be caught earlier.

Bug: catapult:#4521
Change-Id: Ia7cc65592ccc2e8a7d9d86151ae1681d824eb492
Reviewed-on: https://chromium-review.googlesource.com/1109823
Reviewed-by: Charlie Andrews <charliea@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>