[Systrace] Raise error when specified device is not found
Currently Systrace will silently exit when a serial number for a
device that does not exist is specified via either ANDROID_SERIAL or
the -e commandline option.

This change modifies Systrace's behavior to throw an error when the
user requests a specific device that is not connected.

Change-Id: I237c596cafd3252c60814c31a329407b13b0de48
Reviewed-on: https://chromium-review.googlesource.com/1123777
Reviewed-by: Chris Craik <ccraik@google.com>
Commit-Queue: Chris Craik <ccraik@google.com>