[Telemetry] Fix logic in Tab.testGetRendererThreadFromTabId
The test was previously asserting the trivial statement that a list
with a single element on it (which happened to be an iterator) had
exactly one element.

Actually unrolling the iterator showed that it was in fact empty,
probably because the logic to find timeline markers changed at some
point and this broken test failed to detect it.

This CL now:
- Adds the necessary logic to a new thread.IterTimelineMarkers,
- Implements model.FindTimelineMarkers in terms of this new method,
- and fixes the assertion in the test.

Bug: chromium:860297
Change-Id: I84213896c0da97c5e599c6bae342b6af7a3b7527
Reviewed-on: https://chromium-review.googlesource.com/1126306
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>