Fix Bug When No APK is present and Add Relevant Test Cases.
In this CL, we fixed a bug in the long term health tool which would occur when
no APK is present locally. Relevant test cases for the `apk_finder.py` file
is also added to ensure the bug won't occur again, together with the missing
test cases for other parts of `apk_finder.py`.

Bug: chromium:863390
Change-Id: I6ab099b6200a0e4f32b4b1816cc4022676263ce3
Reviewed-on: https://chromium-review.googlesource.com/1183676
Commit-Queue: Leonard Ge <wangge@google.com>
Reviewed-by: Mythri Alle <mythria@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>