Set default_expiration: 20h in v2spa.yaml.
The default cache timeout for static files is 10 minutes.
Increasing this timeout will increase the browser cache hit rate for
vulcanized.html, significantly lowering average load times.

The docs recommend setting this to 1 hour so that, if a bad version is deployed,
users will not need to wait for longer than 1 hour for their browser to fetch a
new version.
https://cloud.google.com/appengine/docs/standard/python/config/appref#static_cache_expiration

20 hours is chosen here so that clients should only need to refetch
vulcanized.html the first time each day that they access v2spa. If a bad version
is deployed, then users will need to either wait a whole day or clear their
cache. I expect users of v2spa to be savvy enough to clear their cache, but we
can re-evaluate this decision if and when such a situation arises.

Bug: catapult:#4461
Change-Id: I9a161fb63ad3f2aa9f40c53cc2f2bf90d1b39004
Reviewed-on: https://chromium-review.googlesource.com/1185287
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>