androidStartupMetric: skip the first start
The first browser start has different behavior for both warm and coldish
starts because of less-controlled factors arriving from installation of
the APK, more system idle time prior to the first run, etc. Analysing
these first starts is more complex because of the said conflating
factors, and also this scenario is less common in the field. Ignore it
for now.

Bug: chromium:881384
Change-Id: I209a2e9eab1ba283837d16e17a5be932fb8bf33b
Reviewed-on: https://chromium-review.googlesource.com/1256882
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>