Fix flaky cp-toast.autoclose-only-last test.
This test was flaky because cp-toast was using timestamps to differentiate between
calls to open(), but the test was calling them in rapid succession, so sometimes the
timestamp didn't change between calls. This CL changes cp-toast to allocate GUIDs by
incrementing a counter to ensure that different calls to open() always get different IDs.

Bug: catapult:#4461
Change-Id: I32f3173e0567230b5f0e53f6b4679d2ccf0c7007
Reviewed-on: https://chromium-review.googlesource.com/1257827
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>