Telemetry: detect telemetry target renderers
In this change we detect the "target" renderer, that Telemetry scripts create
and run JS on, to be used in queueing_duration. Since this will add to
Telemetry-related logic in rendering_metric.html, we also factor out
Telemetry-related logic in a separated helper.

Also, rendererWithLargestPid in chrome_model_helper.html is supposed to do a
similar thing, but in a hackier way. It is not used anywhere. So, we delete it.

Bug: chromium:890757
Change-Id: I83397a7b2f5601a211396ff74e2cc4c0de7ae1ca
Reviewed-on: https://chromium-review.googlesource.com/c/1291889
Commit-Queue: Ehsan Chiniforooshan <chiniforooshan@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>