Integrate test expectation-parsing code into TYP.
As noted in earlier CLs on this bug, this adds support for
the "tagged test list" variant of bit.ly/chromium-test-list-format
to TYP.

Programmatically, this is done by adding support for a single file
argument (-X/--expectations-file) and a list of tags (-x/--tags),
both specified as args to the runner like most other parameters.

The "tags" must be the subset of tags specified in the
file that should be applied to the tests in this run. For example,
if the tags are [ Debug Release ] and [ Mac Win Linux ], then
you might specify `-x Mac -x Debug` for a debug Mac build.

Bug: chromium:835690
Change-Id: I165964e4f4f6403f01669500d0ea90a99e2e8b81
Reviewed-on: https://chromium-review.googlesource.com/c/1171876
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>