[tracing] Check for both type and value presence
Changes the behavior in chart_json_converter.html to check for either
value.type or the actual value fields when decided how to add samples
to the histogram. This fixes samples not being added if the type isn't
explicitly set, but the data is otherwise valid and unambiguous.

Bug: chromium:906685
Change-Id: I4510ca4dfef82ef8f29f997e1e9df2e0d3f6ae54
Reviewed-on: https://chromium-review.googlesource.com/c/1370690
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>