Adds ability to find out if a test is retry on failure
Modifies expectations_parser.TestExpectations.expected_results_for. It
changes the name of the function to expectations_for. Logic is also
modified so that it finds if any of the expectations that match the
inputted test's name has the should_retry_on_failure flag set to True.

Bug: chromium:698902
Change-Id: Ib0f6802a06613de2e6443ee2eb13084dc9069a88
Reviewed-on: https://chromium-review.googlesource.com/c/1474712
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>