Split Histogram.statistics_scalars into statistics_names, GetStatisticScalar.
Currently, the JS implementation of Histogram allows callers to collect the
names of statistics separately from their scalars, but the python implementation
does not. There are situations in results.html where this feature is useful, but
there haven't yet been any such situations for the python implementation.
The new HistogramSet JSON format is moving summaryOptions from a property of
Histogram dicts to a GenericSet Diagnostic containing the statistic names so
that summaryOptions can one day be either removed or significantly simplified.

This CL refactors the python implementation of statistics_scalars to match the
JS implementation.

Bug: chromium:918208
Change-Id: Ib83c0b7b28d8248f7e0c589a93399af8c289dfbc
Reviewed-on: https://chromium-review.googlesource.com/c/1474608
Commit-Queue: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>