telemetry: Add TraceValue.SerializeTraceData().
Add TraceValue.SerializeTraceData() to explicitly trigger serialization
of the trace data. This is part of the CL to process the traces in a
thread-pool (http://crrev.com/c/1474609), but this needs to happen
first so that tests in chromium can be updated for the new API, before
the main CL can re-land.

BUG=chromium:935834

Change-Id: Iedeaf8d68dc1fe218bf44bdaaad2b41443e950f3
Reviewed-on: https://chromium-review.googlesource.com/c/1496612
Auto-Submit: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>