Add dont-require-rooted-device compatibility-mode to Telemetry.
This avoids doing all of the operations which require root access. It
switches the command line file to the version in /data/local/tmp, and
also places the user profile in a directory under /data/local/tmp.

Tested locally by running a subset of the WebGL conformance tests on a
non-rooted Pixel 2 with Chrome Canary:

  ./content/test/gpu/run_gpu_integration_test.py webgl_conformance \
    --browser=android-chrome-canary \
    --compatibility-mode=dont-require-rooted-device \
    --test-filter=gpu_tests.\
    webgl_conformance_integration_test.WebGLConformanceIntegrationTest.\
    WebglConformance_conformance_attribs\*

Bug: chromium:826295
Change-Id: Ifdfbda7697f6336e1732f60cff8f078f276ad6bf
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1490383
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>