[Telemetry/Metrics] Skip complexity of map_runner.py when only running a single trace.
We are hanging somewhere in there, but it doesn't make sense to even use that
code. This skips map_runner.py, whose purpose is to run the traces in a thread pool,
but we are only running one trace, so it makes a pool of 1 job.

Bug: chromium:953371
Change-Id: I8f2143c8b0a09374814925a98ed3241ee682c11c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1586191
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>