telemetry: implements the CpuTemperature wait
This adds an additional wait between runs similar
to the wait battery_utils.LetBatteryCoolToTemperature

The command line argument --wait-for-cpu-temp is
used to specify the temperature at which the device
will wait for between each story. If this arg is not
specified this feature is disabled.

Bug: chromium:383566
Change-Id: I1a5c20238a963865149419f4912e5160cf2a2e12
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1571650
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>