[Telemetry] Refuse to connect to older Chrome versions
We intend to remove legacy code paths to deal with older versions of
Chrome in Telemetry. This should help figure out places in which we may
be inadvertently connecting to older versions by loudly complaining
as soon as the connection is established.

Bug: chromium:962883
Change-Id: Ia70da5458b73e985de84a4169c5fae4767c4b2bb
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1617764
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>