Implement Visually Complete at 2s Metric
We're experimenting with a new kind of loading metric based on last
update time of each pixel. As a first proof of concept of a metric like
this, we want to calculate what portion of a viewport pixels reached
its final value at the 2nd second (the choice of 2 is completely
arbitrary). In this patch, we're considering the whole layerRect to be
updated at each cc::DisplayItemList snapshot. In subsequent CLs, this
algorithm will be made more precise by parsing Skia draw commands and
inferring exactly which pixel values were updated.

Bug: chromium:973838
Change-Id: I89ebdf26bd6ba41e6f703e06a9e404b8174570c3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1652344
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Deepanjan Roy <dproy@chromium.org>
Commit-Queue: Rasha Nasri <rnasri@google.com>