[Devil] Speculative SetEnforce fix
Speculative fix for an issue where SetEnforce was failing due to
insufficient permissions even though it was being run as root. This
appears to only happen if the device becomes temporarily unreachable
during a test, so the suspected issue is that the device is
unexpectedly rebooting or otherwise clearing its root state.

The fix is to clear the DeviceUtils' cache in the event of a
DeviceUnreachableError.

Bug: chromium:979708
Change-Id: Iaa1a2d6ca5ffdcebc2ee526bee81c80e631337a5
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1684227
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>