[Telemetry] Remove "ps" scrapping methods.
Removes all of the following unused methods:
- GetPsOutput
- GetChildPids
- GetCommandLine
- IsApplicationRunning

If we need something similar in the future, we probably would use
psutil python module directly rather than trying to parse the output
of the "ps" command.

Note: browser_backend.GetPid is not removed, as that one is actually
used and does not depend in any of the above.

Bug: chromium:921000
Change-Id: I8d0d190bb9974f07afe9d2cc141c221ed087bab0
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1683955
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>