Speed index implementation based on first paint rects
Measure visual completeness at different points in time by
calculating the area of the viewport that was painted so far compared
to the final total area painted.
AreaSoFar is calculated by summing the areas of the rects that were
painted before that point in time (for now, we're assuming that paint
rects don't overlap).
Use the visual progress array to find the speed index.

Bug: chromium:982025
Change-Id: Idab5b9c4756c7f1036aea1eb99dc892be7491a87
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1691314
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Rasha Nasri <rnasri@google.com>