Pinpoint - Make autocomplete-box validatable
This adds a required property to autocomplete-box and adds an
IronValidatableBehavior so that it works nicely with iron-form. This
way we can use autocomplete-box in a form, add the required property,
and missing inputs will show up red.

This is split off from https://chromium-review.googlesource.com/c/catapult/+/1703521

Bug: chromium:982391
Change-Id: I99d78d68403322b421332337a9e471adc55efc73
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1709750
Reviewed-by: Dean Berris <dberris@chromium.org>
Commit-Queue: Simon Hatch <simonhatch@chromium.org>