Catapult: Add a flag to disable Android tests
This change allows us to eventually turn off the Android tests in case
we determine that we don't want to run those for some reason. The main
reason for disabling these tests would be if we detect from LUCI that
we're only changing files in the Dashboard project (e.g. no changes to
Telemetry).

Those changes in the builder configurations live in a different
repository, so we'll need to make this change first before changing the
confifuration for the try bots.

R=crouleau@chromium.org, simonhatch@chromium.org

Bug: chromium:985223
Change-Id: I0c678217ebf9027115396a146f586eb5b3e9bd5a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1707794
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>