PinpointScheduler: Support job cancellation
Now that we have a scheduler API, we should allow users to cancel jobs
that are either 'Queued' or 'Running'. We expose this in the UI as
elements which show up alongside currently running jobs in the job list
that are eligible for cancellation.

We follow the conventions we have for the APIs used by Pinpoint, but in
the future we should migrate job management into a proper API hosted
through Cloud Endpoints.

In the meantime we support the following cancellation scenarios:

- Users who start a job can cancel their queued or running jobs.
- Administrators can cancel anybody's job.

Future changes should allow us to re-run cancelled jobs like we can let
users re-run previously finished/failed jobs.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:977488
Change-Id: I1548bd5f3b17cde56066c68412a0f15bf3275571
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1710035
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>