Implement core screenshots based speed index
We are given a list of screenshots(as an array of pixel values) and
a list of timestamps that are associated with the screenshots.
To be able to compare how similar two screenshots are, we are going
to extract a color histogram for each screenshot and use the earth
mover's distance to measure visual completeness at different points
in time. And given the visual completeness data we can calculate the
speed index.

Bug: chromium:983638
Change-Id: Ie65123dd5c7288e87d9705203784b247c12a4d7d
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1712446
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Deepanjan Roy <dproy@chromium.org>
Commit-Queue: Rasha Nasri <rnasri@google.com>