[Telemetry] Create an IsLowEnd function for the AndroidPlatformBackend class
This function returns True if an android is a gobo or W6210 model or if it has
a svelte build. This function also removes the android-svelte Typ tag and replaces
it with the android-low-end tag. We are removing the svelte tag because it is not
very well defined. The low-end tag more clearly describes the distinguishing
feature of the kind of devices that use this build.

Bug: chromium:973936
Change-Id: I42a35c58170998e084f9201be94dfef5b18bd9ac
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1725113
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>