Dashboard - Surface errors from Pinpoint more clearly.
Currently Pinpoint doesn't check if a bot configuration exists, throwing
a cryptic exception instead.

Currently the dashboard doesn't surface errors from Pinpoint, because
the request library throws exceptions on anything other than a 200.
This changes it to pass the original contents as a property on the
exception, and pinpoint_service can surface the error to the UI.

Bug: chromium:992532
Change-Id: I6e3d335340d98b8df48f2415be6fe7439a57911d
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1747177
Reviewed-by: Dean Berris <dberris@chromium.org>
Commit-Queue: Simon Hatch <simonhatch@chromium.org>