[Telemetry] Uncaught exceptions should interrupt benchmark
This is normally not an issue since Telemetry's story runner loop does
catch and properly record exceptions raised while running a story.

However failures on tests often raise exceptions which remain uncaught,
and the assert to check that the results object is not finalized while
stories are still running obscures the root cause for why the test
failed.

This CL makes it so uncaught exceptions also get recorded as benchmark
interruptions, and skips over the assert as it's likely that the
exception did happen while running a story.

Bug: chromium:973837
Change-Id: Icc8ef4aabc8bdad28a9cf212d3e190e3c721f4f6
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1786999
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>