[Telemetry] Add system condition tags frozen set to typ expectations module
This set includes all the legal tags that can be used in expectations.config.
We will add a presubmit check which make sure that the expectations file has
all the tags in this set. It also adds the user friendly name for Android
Kitkat.

Bug: chromium:999335
Change-Id: I979cfbebf3f4f1d99617a67e6245156e5e6541ed
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1789691
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>