Devil: default to all devices if --device not specified
If the --device flag is not specified, we should default to all healthy
devices. This was already promised by AddDeviceArguments() (which
appends the --device flag to the ArgumentParser), but was not
implemented correctly (args with action='append' counter intuitively
default to None).

Bug: chromium:1002777
Test: Manual - android_webview/tools/remove_preinstalled_webview.py works as intended
Change-Id: I61a4e41913624ae3413c2a590014cae1b72f8d3f
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1799291
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>