Pinpoint: Pull out benchmark args to top-level
This change pulls out the job arguments relating to benchmark runs and
analysis options from inside the `arguments` dictionary into a
structured property in the Job model. We're doing this to make it easier
to analyse the data in a structured format when we export this to
BigQuery.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:1006658
Change-Id: If01a24f13006606c43e3b14da6dbc54fd65d31be
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1816336
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Commit-Queue: Dean Berris <dberris@chromium.org>