PinpointScheduler: A Validator Implementation
This change implements a validator for the state of a given task graph
which is applied through the Evaluate framework. From this point, we
have a chance to do the following:

- Implement validation as part of the Evaluator implementation(s) so
that all validation can be done in sequence (using the combinators).

- Isolate the validation process of the whole task graph from the
evaluation of what actions to perform.

Because we're only using this in the RunTest implementation, we're
showing how this can be done as an isolated step.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: I3a58ff22751c4fce1d3a4d182f5c7a6bedb6ac39
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1819079
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>