Add SubSysId and Revision as part of a GPU device identification.
On Windows this is a way to distinguish when a GPU is used on
devices from different OEMs. Theie available features and performance
could be very different.

This CL is necessary for wiring these two pieces to telemetry.

This works together with
https://chromium-review.googlesource.com/c/chromium/src/+/1824457

However, the handling of these two new fields are backward
compatible, so this CL can land independently.

Bug: chromium:1006960
Change-Id: I1a2280b81ab7d1f2d6d9e3f1ab5f14a6c52644f2
TEST=gpu_integration_tests
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1825986
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>