[typ] Update artifacts module to handle web tests artifacts
Blinkpy creates artifacts for almost all tests. We can use typ's
artifact module to keep track of these artifacts and then use it
produce a dictionary that can be added to the JSON results. This
CL makes modifications to the artifacts module so that it can keep
track of web test artifacts.

Bug: chromium:982208
Change-Id: I2c4fc624628502793c1047a88219804286c8a2f3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1825270
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>