PinpointScheduler: Port ReadValue to Tasks+Evaluators
This change implements the ReadValue quest into tasks and evaluators,
which is run in the execution engine. This change also refactors the
task creation logic into factory functions which can be configured with
well-defined task options, with a clear dependency graph when creating
the structures which have dependencies.

In this change, we have a tighter coupling between the ReadValue
task+evaluator implementation with the run_test implementation. We're
attempting to consolidate the logic of reading data from the isolates
into a singular type (ReadValueEvaluator).

The `CreateTasks` factory functions allow us to generate subgraphs with
the well-defined couplings. For instance, we can represent the following
graph, with the requisite dependencies:

    ReadValue[ 0 ] -> RunTest(CL[0],  0 ) -> FindIsolate(CL[0])
    ReadValue[ 1 ] -> RunTest(CL[0],  1 ) -> FindIsolate(CL[0])
    ReadValue[...] -> RunTest(CL[0], ...) -> FindIsolate(CL[0])

Or, alternatively:

      ReadValue[0]     ...    ReadValue[N]
          |                       |
          v                       v
    RunTest(CL[0], 0)  ...  RunTest(CL[0], N)
          \                      /
           \                    /
            \                  /
             v                v
             FindIsolate(CL[0])

This lets us track individual "attempts" or runs/samples which depend on
the graph structure.

Future changes will allow us to implement a higher-level 'FindCulprit'
task for bisections, which can populate this graph and depend on the
individual `ReadValue` tasks' results to determine whether to add more
attempts, or whether we have enough information to determine a culprit,
and then introduce additional tasks.

For A/B or try jobs, we can do the attempt scaling for increasing
confidence in the resulting A/B results.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: I770eb7f93d8dbf5d65bd74756e1333664cfab5c4
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1830259
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>