[Telemetry] Move screenshot-on-failure tests to shared_page_state tests
The relevant code involved in screenshot taking is in SharedPageState
class, so it makes more sense for the tests to be there.

Also take the chance to rewrite the last test which was using
json-test-results output format. After this we can remove support for
the format entirely from Telemetry.

Bug: chromium:928275
Change-Id: I20c78797e033ac87dfb6bae95cf2af66908dcd9f
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1860054
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>