[Telemetry] Implement --story=<exact story names> flag.
This provides a way for Pinpoint or other benchmark runners
to provide simple and exact names for stories to run.

This is an alternative to the regex used by the --story-filter
flag, which is rather complicated to use.

Bug: chromium:982027
Change-Id: Iee56d5c81cc2fc3a5c7957bdcbcf983b8220e4a3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1869800
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>