[Telemetry] Fix and update testScreenShotTakenForFailedPage
The "interesting" part of the test didn't actually ever run, because of
a badly written condition which always evaluated to False.

Today, because tab.screenshot_supported is always True [1], we should
always be able to gatter some sort of screnshot in case of failure.

So in this change we fix and update the test under this assumtion.

[1]: https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/internal/backends/chrome_inspector/inspector_backend.py?rcl=9c2e77d9113f9aaff6fe6a44a3439092b956494b&l=151

Bug: chromium:1025750
Change-Id: Ib636eaa5fd1e5b273a97e40314d0ed818963929d
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1922067
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>