Implement a Job Serializer Evaluator
This change defines an evaluator that can take a task graph and turns it
into an update to the job dict when serialized. This evaluator is
independent of any tasks, so we consolidate the logic involved in
extracting the important information from the tasks.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: I7942773914e55046da443e7aed7246b66dee98e3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1868196
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>