Update build steps to allow subsetting tests
This change allows us to run different builders for running just the
Dashboard tests. For now we're it the default that we'll run all the
tests, until we land
https://chromium-review.googlesource.com/c/chromium/tools/build/+/1935583
which allows us to define the properties that will enable us to
configure a try bot just for the Dashboard and Pinpoint.

In a follow-up change, we'll remove the Dashboard tests from the default
set of tests run when other parts of Catapult are changed.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:985223
Change-Id: I35fadc24b1bc955c623b6ac651f8224ee715cb11
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1935278
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>