Fix displaying wrong job when switching between jobs
There are widely misusing of the `dom-if` pattern to prevent the code
execution similar to chromium:1029025. This is only a workaround to fix
most of related bugs by simply check the requirements every time
callback invokes, and it seems greatly increase the usability of
pinpoint's Web UI.

It's not a good solution, but it's the simplest way to make the
current UI work before we can rewrite it.

Bug: chromium:1029412
Change-Id: If2611f678e28e077b2336c86b524d0adc4ece316
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1943731
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>