Pinpoint: Blame only top 2 culprits
When Pinpoint encounters multiple culprits in a bisection range, we
apply a heuristic which will only CC folks from the top two commits with
the largest differences. This limits the amount of potentially CC'ed
users for issues that can be found in really noisy metrics being
bisected through, where the number of potential culprits is non-trivial.

In the future we should improve the algorithm that finds the differences
to consider globally significant changes only, but in the meantime we'll
apply this heuristic to reduce the noise Pinpoint can generate.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1021523
Change-Id: I94f1df550a00fdf91339083f920c1c396616dbb5
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1947847
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Andrew Bennetts <abennetts@google.com>