Parallelize Breakpad symbol dumping
Changes _GenerateBreakpadSymbols to run each instance of
generate_breakpad_symbols in parallel instead of serially.

This is meant to address the fact that dumping symbols for component
builds takes a very long time, particularly on Mac. Each call to
generate_breakpad_symbols takes a reasonable amount of time, but there
are a large number of binaries that it needs to run on in component
builds.

Running everything in parallel results in a significant speedup on a
Macbook Pro 2018. Running Chromium's Telemetry minidump tests without
this change takes 884 seconds, while they only take 360 seconds with
this change.

Bug: chromium:1027523
Change-Id: Icc06f33bb0077cb4e00c84602767502ddb54b761
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1990512
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>