[Benchmarking] Add timeout for TakeScreenshot on Mac
We noticed some Mac bots wait for 6 hours after certain tests failed.
When looking at the logs, we found that one of the step after failure
is taking screenshot and the 'screencapture' process remains active
at the point of the test timed out.

In this case, we added a mechanism to introduce a much smaller timeout
as 15 seconds when launching the screencapture process. It utilize
existing WaitFor() function in py_utils.

Bug: chromium:1036447
Change-Id: I2a067e75b53e55fd33b4deb8cd53ec77910c75da
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1990354
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>