Add support for diagnostics and unit conversion.
The tests need to be able to write the unit test story diagnostic
into the histogram, so adding support for that. I have essentially the
same interface as histogram.py.

We have a couple other options here: we could either deviate from the
histogram.py interface and just add a set_story method that creates
the diagnostic behind the scenes. This has the advantage of not leaking
the proto to users of histogram.h, which makes the proto easier to
change. We could also create wrapping types for Diagnostic, Unit and
GenericSet which hides the proto. That will add quite a bit of
boilerplate code, but will hide the proto.

Bug: chromium:1029452
Change-Id: Ibbf840cb98c6566c3063b1778b014479f5667e11
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2002509
Commit-Queue: Patrik Höglund <phoglund@chromium.org>
Reviewed-by: Dean Berris <dberris@chromium.org>