Pinpoint: Fix results2 for new executions
This change adds test coverage for the results2 processing code,
ensuring that we're actually testing the execution dispatch routines.
This gets us higher confidence that we won't be breaking this in the
future as we make more changes to the code related to results
processing.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1056658
Change-Id: Ie2f5fa6e0e3f90797b0943341c9b2cae20a35de3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2078499
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Andrew Bennetts <abennetts@google.com>