Implement proto writing in the dashboard.
This rounds out the proto implementation. It enables for instance
the WebRTC catapult_uploader.py code to send protos directly to the
dashboard - today it reads protos but sends it to the dashboard as
JSON. This is OK but the proto wire format is more efficient.

This is also needed for a WebRTC python test which writes results
from python.

Bug: chromium:1029452
Change-Id: Ib94587f4bb86ce9121da46dce1186fc503bfea82
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2111217
Commit-Queue: Patrik Höglund <phoglund@chromium.org>
Reviewed-by: Dean Berris <dberris@chromium.org>