rendering: Add a missing param.
It looks like we generate the various cpu-time metrics twice
because of a missing input param. So fix that. This is not
expected to affect the output results, but perhaps can bring
down the runtime a little bit.

BUG=chromium:1058155

Change-Id: If9116c0eb19078e53dab4ac1ef570ce04d1f30a3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2143453
Auto-Submit: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Behdad Bakhshinategh <behdadb@chromium.org>
Reviewed-by: Deep Roy <dproy@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>