Cleanup of how Job generates issue tracker updates.
The motivation for this is to make it easier to improve the issue
tracker updates posted by Pinpoint, but significantly changing the
content of those updates is not in scope for this change.

Mostly this change is a refactoring:

* extracting a bunch of logic for generating
issue tracker updates about found differences from job.py into a
new module.
* keeping state in simple explicit objects, rather than having many
variables and function parameters and return values with implicit
invariants between them.
* rename various functions to have clearer names, and add many more
comments.

However there is a small opportunistic bug fix due to simplifying
the logic in _ComputePostOwnerSheriffCCList (now called
_PeopleToNotify), so this adds a unit test for that.

Bug: chromium:1076756
Change-Id: I2cbd8388cb72ec8e13a5c319a63ccd9c77cce5eb
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2172930
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Dean Berris <dberris@chromium.org>