Fix Android symbolization race condition
Fixes a race condition in Android minidump symbolization when
symbolizing Monochrome minidumps. Previously, we were simply returning
all possible symbol binary locations and assuming that the correct
dumped symbols would be used. However, in cases of mixed bitness such
as with Monochrome, it's possible to have valid symbols in multiple
locations. This, combined with the fact that non-official builds
typically have null debug IDs for libchrome.so/libmonochrome.so, meant
that the wrong dumped symbols could be used depending on which of the
valid locations was dumped last.

The fix is to check for a directory's viability for symbols for a
particular minidump, and only return symbol binaries in a single
directory.

Bug: chromium:1056235
Change-Id: I8d89be06ce63502f421e54a1e28c62e85c98b32e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2188954
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>