[Telemetry] Switch arch matching to regex
Switches the way we match and extract the processor architecture in
android_minidump_symbolizer from "in/split" to a regular expression.

The current approach is flakily hitting an issue where the split is
returning a list with a single element, implying that there might
occasionally be a line in the minidump_dump output that contains
"processor_architecture" but is not the line we want. The use of a
regular expression should prevent matching against that.

Bug: chromium:1056235
Change-Id: I9e0bc41edd42adf51784e2a6eb65817db6704863
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2191171
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>