Devil: add GetApplicationTargetSdk helper
This adds device.GetApplicationTargetSdk() to fetch the targetSdkVersion
of an application installed on the device. Like
apk_helper.GetTargetSdkVersion(), this will return the value as a string
and will convert pre-release SDK versions to the platform codename.

This is a prerequisite to more-detailed error messages for switching
WebView provider.

Bug: chromium:1034860
Test: ./devil/bin/run_py_tests
Change-Id: I9bd92120199eaa27ea82f0e58db1d0e811412715
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2190248
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>