auto-triage: Support projects in anomalies
This change touches both the Dashboard and Pinpoint in that we're
allowing calls that previously assumed bug id's stood alone to now also
support taking a project argument. This change lets us merge issues
across and within projects.

In this change we are also adding an index for the project_id which is a
new field in the anomaly model.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1082030
Change-Id: Ife7dee8d7f2f1384238dd4ad2484f6ce82e5d8d9
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2198879
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>