auto-triage: File issues on specific projects
This change implements support in the Dashboard for auto-triage to file
issues that match subscriptions defining multiple projects. The change
involves two major pieces:

1. We're creating one group per Monorail project for matching
auto-triage configurations. This allows us to differentiate alert groups
for different projects which match the subscription configuration. This
allows folks from different projects to file individual issues for the
same alert group for tracking the fixes/investigations independently.

2. We're preparing the Anomaly model to support multiple issues with
projects associated in the model. This is only in preparation until we
find a way of representing this in the Dashboard UI, and until we can
ensure that we can transactionally update the list of issues to avoid
eventual consistency conflicts.

We're also taking some opportunity to refactor the fake issue tracker so
that it has value/object semantics instead of being a global.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1082032
Change-Id: I176dc71cfd7e73188d5604a6c84181dc715667c2
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2206622
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>