auto-bisect: Validate more input parameters
This change ensures that we validate the inputs for when new Pinpoint
jobs are started. We know that target and benchmark at least must not be
empty strings even if they were provided as inputs to job creation.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1086315
Change-Id: I42e4f161b6a534155eeedca87f4fcc3570a8dab7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2214970
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>