auto-bisect: Skip bisection for single-rev ranges
This change checks whether the commit range is a single commit or
whether there's only one difference in the range. In those cases we:

- Find the author of the revision, and assign it directly.

- Add the Chromeperf-Auto-Assigned label to differentiate the changes
  that skip bisection from those that are actually bisected with
  Pinpoint.

We are re-using as much of the current bug-filing mechanism in the
file_bug module here, to reduce the need to reimplement all this
functionality.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1087803
Change-Id: I9ceb49152c27f2d509a3b4b30aceb751294ca3fa
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2224082
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>