Remove the pipeline metrics and visualization charts
This CL removes the pipeline:* metrics in favor of the UMA metrics for
the same stages. The entire visualization container that holds the
breakdown charts associated with these metrics (along with the
"Visualize" button to toggle the visualization container) in
results.html are also removed.

Bug: chromium:1078160
Change-Id: Iaff9faa570d4a5ffd0b227ad1670ff2100fa4759
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2212422
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Reviewed-by: Deep Roy <dproy@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>