Typ: Go back to using io.StringIO implementation of _TeedStream
Revert: "Reland [Typ] Use a file to store sub process output"
This reverts commit fd813d1131850484e965a0279718be10f8aed48f.

The current implementation for _TeedStream is complex and may be
unmaintainable. I even noticed a difference in behavior between Mac
and other platforms when implementing this change which is summed
up in the following source comment in _TeedStream

"# on mac python stream object does not account for captured bytes
 # added when another host object captures the output and turns on
 # passthrough"

It is not worth maintaining this implementation just to remove
subprocess output from tests that spawn typ subprocesses. It would
be easier to modify these tests so that they don't run typ in
a subprocess or just implement a quiet mode for typ and have tests
run typ in quiet mode instead.

The potential other problems that were caused by this change include
typ hanging when running with multiple processes on Mac bots and
missing log output after the Browser restarts during GPU test runs on
Mac.

Bug: chromium:1047574
Bug: chromium:960833
Bug: chromium:960606
Bug: chromium:952105
Bug: chromium:967922
Change-Id: Ibf5ab8863802ddf1bb96ced53a84261e62231df3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2254494
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>