[rendering] Remove breakdowns for cpu-time metrics.
The total_rendering_all_* and total_all_* metrics (for cpu-time per
frame and tasks per-frame) get breakdowns for the contributing threads.
This is not particularly useful, since (1) the breakdowns are already
reported in their own histograms, and (2) the UI is not obvious enough
that most developers do not use the breakdowns anyway. Therefore, remove
breakdowns for these metrics.

BUG=chromium:1073148

Change-Id: I7bc44ffb2fe9092068220af90899713bc2afd909
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2269869
Reviewed-by: Deep Roy <dproy@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>