[v8][wasm] Introduce metric for baseline streaming compilation
This CL introduces a metric for baseline compilation of streaming
compilation in WebAssembly. The metric measures the wall time, and is
defined as the time from when the first bytes were received until
baseline compilation is finished.

This metric may include the time the wasm compiler is waiting for bytes
to arrive, in the case that compilation speed is faster than download
speed. However, when I looked at the trace locally, even on my 48 core
machine this was not the case.

R=mythria@chromium.org

Bug: chromium:1084929
Change-Id: I3428a9a8f73be13b8c6fea32b48368c62681808a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2277740
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Mythri Alle <mythria@chromium.org>
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>